Ensure data exists before attempting type translation
If the case where Popen produces a non-zero return code, and
none_on_ignored_error is True, then data will be set to None.
For translating the results into the desired type, when
split_lines is true, then len(data) is checked which is not
valid for a NoneType object.

Fix is to ensure data is not None before attempting type
translation.

Testing Done:
1) Replicated failure from https://groups.google.com/forum/#!topic/reviewboard/CYTxQ6ADe_M and confirmed the code mod resolves the issue.
2) Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7183/