More robust parsing of SVN client version
A user of SlikSvn reported an error when parsing the svn client
version string.  For SlikSvn the version string is of the form:
  1.6.1-SlikSvn-tag-1.6.1
which is more complicated than the version string produced by
other svn clients where the format is:
  1.6.1

String parsing is updated to extract exactly the first three
numbers in the version string.  If less than three numbers are
present, which is an unlikely occurrence based on historical
SVN version numbers (https://subversion.apache.org/docs/release-notes/release-history.html)
then a warning is logged and a default version number of
0.0.0 is assumed.

Testing Done:
1) Tested with simple version string produced by stock subversion client in CentOS 7.x.
2) Stepped through code and manually tested with more complicated SlikSvn version string.
3) Ran unit tests.

Bugs closed: 3834

Reviewed at https://reviews.reviewboard.org/r/7182/