Add --svn-prompt-password option
The `--svn-prompt-password` option will prompt the user for their SVN
password the first time an SVN command is run. This is useful so as the
`--svn-password` option will leave the user's password visible in plain
text on their console and in their history, which can be a security
concern on shared systems.

Testing Done:
Successfully used the `--svn-prompt-password` option when using an SVN
repository with the remote repository over HTTP. The password was
prompted for and the command worked.

Reviewed at https://reviews.reviewboard.org/r/7206/