Always run svn log --xml non-interactively
We cannot run `svn log --xml` interactively because the authentication
prompts are intermixed with the `svn log` output. This results in
exceptions when attempting to parse the XML and leads to cryptic error
messages.

This patch replaces all `svn log --xml` usage with non-interactive usage
and checks for a specific error code to determine if there was an
authentication error. Authentication errors are written to the console
with directions on how to rectify them using the SVN authenticaiton
command line options.

Testing Done:
Ran unit tests.

Ran `rbt diff` in an SVN repository with specific revisions with a
remote repository over HTTP with authentication. RBTools exited with
an error that prompted me to enter my SVN credentials on the
command-line.

Reviewed at https://reviews.reviewboard.org/r/7205/