Override session cookies with username flag
Previously, when the `--username` flag was provided to RBTools and a
session cookie already existed, RBTools would prompt for the password
but never use it (because of the session cookie). We now clear the
session cookie when the `--username` flag is given.

Testing Done:
Ran unit tests.

Attempted to post a review request with the `--username` parameter and
a bogus password; the review request was not created. It was created
when the correct password was used.

Bugs closed: 3858

Reviewed at https://reviews.reviewboard.org/r/7289/