Allow the HTTP cache to be disabled or created in any location
The HTTP cache for the API can now be disabled via the
`--disable-caching` command line argument. It can also be moved to an
arbitrary location with `--cache-location FILE` or created in memory
with `--in-memory-cache`. This behaviour is also available to consumers
of the API via the `RBClient`. The `allow_caching`, `cache_location`,
and `in_memory_cache` key word arguments can be used for this.

Testing Done:
Unit tests passed.

Ran the following commands:

 - `rbt status --debug --disable-cache`
 - `rbt status --debug --in-memory-cache`
 - `rbt status --debug --cache-location /tmp/cache`

The debugging information indicated the correct behaviour was taking
place in each instance.

Reviewed at https://reviews.reviewboard.org/r/7293/