Add support for an in-memory cookie store
There is now support for an in-memory cookie store. This feature can be
enabled via the `--in-memory-cookies` command-line option. For
consumers of the API, passing `in_memory_cookies=True` to the `RBClient`
constructor will enable this behaviour as well.

Testing Done:
Ran unit tests.

Updated this review request with the `--in-memory-cookies` flag. I was
prompted for my username and password. When updating without the flag,
I am not prompted.

Reviewed at https://reviews.reviewboard.org/r/7295/