Raise a more-useful error when failing to determine the repository.
In the case where the repository isn't correctly identified, we'd attempt
posting while passing ?repository=None to the API request. This then returns an
ugly error from the server. This change adds a check to see if the detected
repository is None, and bails out with a readable error message.

Trivial change.

Fixes bug 3687.