Print the reason for ServerInterfaceErrors.
There are many possible reasons for not being able to communicate with a Review
Board server, and now that Python is validating SSL certificates by default, it
may be very confusing if that's the one that bites people. This change adds the
exception message to the error output. Trivial change.

Fixes bug 3840.