Modernize the RBTools README.
Our old README was a ReST-formatted file with some basic information,
but it didn't render as anything but plain text on GitHub, and didn't
really provide enough help on getting going with RBTools.

This new README fleshes that out, converting it to Markdown and
providing a lot of links to useful pages. There's an introduction to
RBTools, download instructions, information on the API, links to docs,
contributor/testing information, and the standard set of support/Happy
Users/bug reporting we have in the Review Board README.

Testing Done:
Read through the new README in Marked, and tested all the links.

Reviewed at https://reviews.reviewboard.org/r/7385/