Defer checking git version until it is needed
Previously, the `gi`t version was checked in the initialization of the
`GitClient` SCM Client. However, with systems that do not have `git`
installed, this would result in an error being printed that `git` could
not be found. We now defer determining the installed `git` version
until we need it, which occurs after we are sure that `git` is
installed.

Testing Done:
Ran unit tests.

Removed `git` from my system and ran `rbt status`. Without this patch,
an error message was printed. With it applied, there was no error.

Reviewed at https://reviews.reviewboard.org/r/7405/