Use the submitter name and email and not resource in land
Previously, when landing a review request, the author would be the
actual resource corresponding to the submitter. Now this has been
recitified to use the name and email address of the submitter instead.

Testing Done:
Ran unit tests.

Landed a review request and saw that the author was correct.

Reviewed at https://reviews.reviewboard.org/r/7409/