Fix plumbing of change number when using --diff-filename.
A previous change of mine fixed an exception when trying to use
--diff-filename, but didn't fix an issue where we had regressed the use of
this command-line argument with systems that use change numbers (such as
perforce). This change fixes the plumbing so that we set the change number
correctly even if the user (for some reason) is choosing the create the diff
themselves.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7501/