Allow 'rbt' to bypass external authentication with prepared extra cookies file.
On systems that have additional external authentication, such as SSO, it's
useful to be able to pre-store additional cookies. This change adds a new
`--ext-auth-cookies` command-line parameter to pass this in.

Testing Done:
Used this in production with a server that needs extra authentication.

Reviewed at https://reviews.reviewboard.org/r/7531/