Check for GNU Diff when using a TFS repository.
Now that TFS requires GNU diff, we need to make sure that it's
installed, and present a helpful error message if it's not. We already
do this for a number of other SCMClients, like SVN and Perforce, and
have a handy utility method for this. This uses that method to do the
check and print the appropriate information.

Testing Done:
Set my path so that GNU diff could not be found but a fake tfs could.
Posted a change and saw the error. Added GNU diff to the path, and saw
it got past the check.

Reviewed at https://reviews.reviewboard.org/r/7615/