Add Bash and Zsh auto-completions.
Bash and Zsh are Unix shells that have auto-completion features.

The command `rbt setup-completion` allows the user to install
completions for Zsh or Bash via the command. However, as Bash
and Zsh are in protected directories that the user cannot
write to, sudo is needed to install the scripts. A manual
installation method is also available in the documentation.

Testing Done:
Ran unit tests for rbt.

Successfully installed completion for Bash and Zsh on OSX.

Successfully installed completion for Bash and Zsh on Ubuntu.

Successfully built documentation.

Reviewed at https://reviews.reviewboard.org/r/7668/