Enables the option for color to be displayed for warnings and errors.
RBtools currently does not have the option for color to be displayed for warnings and errors.
I have changed the command-line UI, so that it defaults to color for warnings and errors, but
allows the user to disable the feature if they do not wish to use it.

Testing Done:
I added logging.warning, logging.critical, and used a command that would give an error to see
if color would be displayed. After that I used the --no-color command to turn off the color off.
Also, I ran the unit tests and it passed them.

Reviewed at https://reviews.reviewboard.org/r/7704/