Fix logging when skipping symlinks in perforce.
When posting a perforce changelist that included a deleted symbolic link,
`rbt post` would die with a TypeError due to a missing argument in the format
string. The problem here was that we were missing the version in the logging
call. Trivial change.

Testing Done:
Tested with the same changelist locally. Now the warning is correctly shown:
**WARNING: Skipping file //depot/filename#1: "//depot/filename#1" is a symlink**

Bugs closed: 4037

Reviewed at https://reviews.reviewboard.org/r/7801/