Add support for publishing with a message
Both `rbt post` and `rbt publish` can now publish with a change
description. This change description supports markdown via the
`--markdown` flag used for summaries and descriptions.

Providing a change description before the review request is public will
not result in an error; instead the option will be ignored and an
informative message will be displayed to the user.

The `publish` command now also only fetches the required links and
fields when publishing to minimize the amount of data that is
transmitted.

Testing Done:
- Published with a message.
- Verified that `rbt publish -m` provided an error message when the
  review request was not public.

Reviewed at https://reviews.reviewboard.org/r/7763/