Don't try asking for username/password when input is not a terminal
This fixes a mysterious "hang" when running rbt using native Windows Python in
a Cygwin terminal emulator such as MinTTY: in this case, password entry doesn't
work correctly and even the prompt is not displayed.

Testing Done:
Ran it from a "normal" DOS window and from MinTTY and got the expected prompt
in the former case and the expected error in the latter.

Reviewed at https://reviews.reviewboard.org/r/7956/