Add support for "branch" operations in TFS.
This change, based on one by Brandon Martin, adds support for the "branch"
operation in TFS, which is essentially a file copy. This can be combined with
edit operations in-place. This will find the original filename and add the
correct metadata to the special header for the server side to know that the
operation was a file copy.

Testing done:
- Did branch and branch+edit operations. Saw that `rbt diff` and `rbt post`
  resulted in what I'd expect.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8033/