Don't use filesystem encoding for interpreting XML from TFS.
A few of the commands we run return XML results, which will always be utf-8,
regardless of what the filesystem encoding is. In this case, we were decoding
to mbcs, which would then fail when trying to interpret some special
characters.

Testing done:
- Ran unit tests.
- Posted a few changes, including one whose parent had unicode in the commit
  message (which is what triggered the bug). Saw that everything worked as
  expected.

Reviewed at https://reviews.reviewboard.org/r/8041/