Make `rbt setup-repo` do fuzzy matching better.
The previous implementation of setup-repo would match one repository per
returned page of results (on release-0.7.x) and one repository anywhere (on
master). If that repository wasn't the desired one, it wasn't possible to
select the right one.

This change increases the number of matches to 4, and sets the similarity
cutoff to 0.4.

Set number of matches to four and cutoff to 0.4

Testing Done:
- Verified behavior of `get_close_matches`.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8043/