Fix a couple bugs when updating Review Requests for Perforce.
There were a couple bugs that were causing problems with Perforce and updating
review requests for changelists. When -u was passed, there was a unicode
mismatch bug that was causing the difflib comparison to always fail, resulting
in no matches.

The second problem was when -u was not passed, there was a bad "update" call
being made (which doesn't actually need to be made). This caused an
AttributeError which had no detail to it, resulting in one of our fantastic
"CRITICAL:" error messages.

Testing done:
Posted a review request with Perforce. Made some changes to the code and
changeset description, and then updated it using both -u (where I was prompted
to choose which review request) and without.

Reviewed at https://reviews.reviewboard.org/r/8076/