Allow 'null' client roots
Windows perforce clients are allowed to specify the client root as 'null'. If
this is the case don't attempt to determine that the current directory is below
the client root because the check will fail. With a 'null' client root all
paths are below the client root.

Reviewed at https://reviews.reviewboard.org/r/8126/