Fix errors running `rbt login` without a server specified.
The `rbt login` command doesn't pass in a tool when looking for a server, which
meant that if a server wasn't specified via the command-line or the config
file, users would get an ugly traceback. This change checks to make sure that
tool is not `None` before calling methods on it, which results in a nice error
message.

Testing done:
Ran `rbt login` outside of a repository with no server specified. Saw the error.

Reviewed at https://reviews.reviewboard.org/r/8154/