Don't use -u in git diff-tree
When generating diffs for SVN and Perforce with excluded files, we would
pass `-u` to `git diff-tree`, which generates diff output instead of the
list of changed files and break the parsing. Now, we remove the `-u`
flag from the `git diff-tree` command when generating the excluded files
list for SVN or Perforce repositories.

Reviewed at https://reviews.reviewboard.org/r/8222/