Fix looking up Subversion repositories by UUID.
In RBTools 0.7.6, we introduced new helpers for iterating over pages of
API list resources, and switched the SVN repository lookups to use it.
Unfortunately, switching form a list to a generator meant that we were
no longer able to loop over the requests a second time for the UUID
checks.

We now store a cached list of the repositories we iterate through, so
that we can loop over them for the UUID step (if needed). We do this
instead of immediately converting the generator to a list in order to
reduce lookups against the API.

Unit tests were added to ensure the conditions all work and won't
regress again.

Testing Done:
Tested against a list of repositories and verified that both loops
were able to operate on the entire list of repositories from the API.

Reviewed at https://reviews.reviewboard.org/r/8226/