Split --tracking-branch option into new group; pass it to rbt land
The `--tracking-branch` option is required for determining the upstream
branch for `rbt land`, but it was not passed in due to it being in the
`diff_options` group. It is now in its own option group,
`branch_options`, which is passed to all commands that accepted
`diff_options` as well as `rbt land.` This fixes an issue where
repositories that don't use the default (e.g., `origin/master` for git)
tracking branch would try to push to a non-existant branch.

Testing Done:
With this patch applied, I am able to do `rbt land --push` when the
upstream branch is not `origin/master`.

Bugs closed: 4161

Reviewed at https://reviews.reviewboard.org/r/8337/