Add missing --tfs-shelveset-owner command-line flag.
The implementation of the shelveset code theoretically allowed users to specify
the shelveset owner (in the case that the owner was not the same as the one who
had created the current workdir), but I had neglected to add this file to that
git commit, and it got lost in a stash. This change adds it back.

Testing done:
Posted a shelveset owned by a different TFS user than my current workdir.

Reviewed at https://reviews.reviewboard.org/r/8353/