Make `rbt install` work correctly on Windows.
Unfortunately, the technique of using `NamedTemporaryFile` in a context manager
and yielding the filename back, then reopening it from elsewhere in the code
caused various permission errors when running on Windows.

This change rearranges the code a bit to make sure we flush and close the file,
and then handles cleanup ourselves.

Testing done:
Used it on OS X, and a customer reports success running on Windows.

Reviewed at https://reviews.reviewboard.org/r/8354/