Fix TFS client to read in the generated diff as bytes.
The execution of the `diff' program from the TFS client was reading in the
output as a `unicode` object, which worked a lot of the time, but was causing
problems in a few cases. This manifested itself most often when doing a diff of
a UTF-8 file which contains a BOM. These files are technically well-formed, but
they're rare in practice because UTF-8 has no need for BOMs. Somewhere in
reading in the output, the BOM in the diff was being converted to a series of
three bytes that weren't a UTF-8 BOM, and so trying to apply the patch later
would fail.

This also would manifest itself when trying to redirect the output of `rbt
diff` to a file.

Testing done:
Successfully did a diff of a UTF-8 file containing a BOM and Review Board was
able to apply the patch.

Reviewed at https://reviews.reviewboard.org/r/8561/