Add release notes for RBTools 0.7.8.
This covers the work done on the installers for Windows and macOS, as
well as the Unicode fixes and improvements to rbt setup-repo. It's a
smaller release, but an important one from an installer standpoint.

Testing Done:
Built the docs and read through them.

Checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/8588/