Do not always overwrite --guess-summary and --guess-description with -g
`rbt post` would always overwrite the value of `--guess-summary` and
`--guess-description` with the value of `--guess-fields`, whether or
not it was actually provided; when `-g` is not provided, all three
flags will default to `auto`.

Now we check for the presence of these fields and only overwrite
`--guess-summary` and `--guess-description` when they are not provided
and `--guess-fields` is. Otherwise, when one of `--guess-summary` or
`--guess-description` is not provided, it will be set to the value of
`--guess-fields`.

Testing Done:
Ran `rbt post` with different values of `--guess-fields`,
`--guess-summary`, and `--guess-description` and saw the expected
behaviour.

Reviewed at https://reviews.reviewboard.org/r/8610/