Add user-configurable option to suppress git, mercurial, and perforce warnings.
Prior to this change, it was not possible for the user to suppress warnings
from git, mercurial, and perforce. This could result in a large number of
warning messages output to the log which can be cumbersome and fill up the
console window.

This change allows the user to suppress warnings from rbtools/clients/git.py,
rbtools/clients/mercurial.py, and rbtools/clients/perforce.py by setting
SUPPRESS_CLIENT_WARNINGS in .reviewboardrc to True.

Testing Done:
Checked that the warning message in git.py is suppressed when the
SUPPRESS_CLIENT_WARNINGS is set to True.

Bugs closed: 4041

Reviewed at https://reviews.reviewboard.org/r/8638/